# src/modules/online/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Online low-code management API
  version: 1.0.0-metadata
paths:
...
  /api/v1/admin/online/definitions/{code}:
    get:
      summary: Get an Online Definition, its current Draft Revision, and releases
      security: [{ bearerAuth: [] }]
      x-permission: infra:online-definition:query
    put:
      summary: Update Definition metadata with optimistic locking
      security: [{ bearerAuth: [] }]
      x-permission: infra:online-definition:update
...
  /api/v1/admin/online/definitions/{code}/schema-plans:
    get:
      summary: List tenant-scoped semantic schema plans; no executable SQL is exposed
      security: [{ bearerAuth: [] }]
      x-permission: infra:online-definition:query
    post:
      summary: Compile and persist a semantic schema plan without applying DDL
      security: [{ bearerAuth: [] }]
      x-permission: infra:online-definition:migrate
...